requirements.txt
requests == 2.28.1
bs4 == 0.0.1
plotly == 5.10.0
fpdf2 == 2.6.0
pandas == 1.5.0
matplotlib == 3.6.0